% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/REPPlabShiny.R
\name{print.epplabshiny}
\alias{print.epplabshiny}
\title{Prints the 'REPPlabShiny' Results}
\usage{
\method{print}{epplabshiny}(x, ...)
}
\arguments{
\item{x}{an object of class epplabshiny.}

\item{...}{further arguments passed to or from other methods.}
}
\description{
Prints some information about objects of class \code{epplabshiny}, typically the result of a call to \code{\link{REPPlabShiny}}. Printed is only a table with the names, indices and algorithms used for calculated \code{EPPlab} object, not the whole content of the object.
}
\examples{
if(interactive()){
 data(ReliabilityData)
 Repplablistshiny <- REPPlabShiny(ReliabilityData)
 print(Repplablistshiny)
 }
}
\seealso{
\code{\link[REPPlabShiny]{REPPlabShiny}}
}
\keyword{print}
